Create generic treatment workflow page to wrap genetic and compound workflows in a common page
This way the compound now also has the plot position settings implemented
There was a huge commonality and in the prospect of having such a third workflow coming it makes sense to not keep duplicating code and having to copy changes between different pages
